+++ graphics/dao_pathcut.h
@@ -114,13 +114,13 @@ struct DaoxPathFragment
 	DaoxPoint  P1;
 	DaoxPoint  P2;
 
-	double  left;
-	double  bottom;
-	double  right;
-	double  top;
+	float  left;
+	float  bottom;
+	float  right;
+	float  top;
 
-	double  start;  /* parametric start location in the original segment; */
-	double  end;    /* parametric   end location in the original segment; */
+	float  start;  /* parametric start location in the original segment; */
+	float  end;    /* parametric   end location in the original segment; */
 
 	DaoxPathEdge  *edge;
 
@@ -135,9 +135,9 @@ struct DaoxQuadNode
 {
 	int    depth;
 	int    count;
-	double  left;
-	double  bottom;
-	double  width;
+	float  left;
+	float  bottom;
+	float  width;
 
 	DaoxQuadNode  *NW;
 	DaoxQuadNode  *NE;
@@ -150,7 +150,7 @@ struct DaoxQuadNode
 DaoxQuadNode* DaoxQuadNode_New();
 void DaoxQuadNode_Delete( DaoxQuadNode *self );
 
-void DaoxQuadNode_Set( DaoxQuadNode *self, int depth, double left, double bottom, double width );
+void DaoxQuadNode_Set( DaoxQuadNode *self, int depth, float left, float bottom, float width );
 
 
 
@@ -167,7 +167,7 @@ DaoxQuadTree* DaoxQuadTree_New();
 void DaoxQuadTree_Delete( DaoxQuadTree *self );
 
 void DaoxQuadTree_Reset( DaoxQuadTree *self );
-void DaoxQuadTree_Set( DaoxQuadTree *self, double left, double bottom, double width );
+void DaoxQuadTree_Set( DaoxQuadTree *self, float left, float bottom, float width );
 void DaoxQuadTree_InsertEdge( DaoxQuadTree *self, DaoxPathEdge *edge );
 
 DaoxPathFragment* DaoxQuadTree_NewPathSegment( DaoxQuadTree *self );
